Cambios en notas pendientes Pt2

+++ ZorzalCacao/Models/FermentacionesDetalles.cs
@@ -10,7 +10,7 @@ public class FermentacionesDetalles
     public int DetalleId { get; set; }
     public int FermentacionId { get; set; }
     public int RemocionId { get; set; }
-    public int Cantidad { get; set; }
+    public double Cantidad { get; set; }
     public double Temperatura { get; set; }
     public DateTime FechaFermentacion { get; set; } = DateTime.Now;
 
@@ -19,7 +19,7 @@ public class FermentacionesDetalles
 
     [ForeignKey(nameof(RemocionId))]
     public virtual Remociones Remocion { get; set; }
-    // empleado
+
     public string EmpleadoId { get; set; }
     public ApplicationUser Empleado { get; set; }
 }
